# spring-debezium-db-kafka-producer/src/main/java/com/learning/cdc/debezium/db/kafka/impl/entity/Product.java
package com.learning.cdc.debezium.db.kafka.impl.entity;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.math.BigDecimal;

@Entity
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class Product {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;

    @NotBlank(message = "must not be blank")
    @Column(nullable = false)
    private String name;

    @NotBlank(message = "must not be blank")
    @Column(nullable = false)
    private String description;

    @NotNull(message = "must not be null")
    @DecimalMin(value = "0.0", inclusive = true, message = "must be greater than or equal to 0")
    @Column(nullable = false)
    private BigDecimal price;

    @NotNull(message = "must not be null")
    @Min(value = 0, message = "must be greater than or equal to 0")
    @Column(nullable = false)
    private Integer stock;
}